pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-starter-python"
version = "1.0.0"
description = "Simple voice AI assistant built with LiveKit Agents for Python"
requires-python = ">=3.10"

dependencies = [
    "livekit-agents[openai,silero,turn-detector]~=1.2",
    "livekit-plugins-deepgram>=1.3.6",
    "livekit-plugins-google>=1.3.6",
    "livekit-plugins-openai>=0.9.0",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv",
    "torch>=2.8.0",
    "numpy>=2.0.2",
    "einops>=0.8.1",
    "einx>=0.3.0",
    "omegaconf>=2.3.0",
    "packaging>=25.0",
    "safetensors>=0.7.0",
    "soundfile>=0.13.1",
    "soxr>=1.0.0",
    "torchaudio>=2.8.0",
    "transformers>=4.57.1",
    "gradio>=3.36.1",
    "huggingface-hub>=0.36.0",
    "voxcpm>=1.5.0",
    "modelscope>=1.15.0",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
